Unify class names
