+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/main/java/org/apache/shardingsphere/spring/namespace/handler/mode/ClusterPersistRepositoryNamespaceHandler.java
@@ -22,7 +22,7 @@ import org.apache.shardingsphere.spring.namespace.tag.mode.ClusterPersistReposit
 import org.springframework.beans.factory.xml.NamespaceHandlerSupport;
 
 /**
- * Cluster persist repository namespace handler.
+ * Spring namespace handler for cluster persist repository.
  */
 public final class ClusterPersistRepositoryNamespaceHandler extends NamespaceHandlerSupport {
