+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/main/resources/META-INF/spring.handlers
@@ -16,5 +16,5 @@
 #
 
 http\://shardingsphere.apache.org/schema/shardingsphere/datasource=org.apache.shardingsphere.spring.namespace.handler.DataSourceNamespaceHandler
-http\://shardingsphere.apache.org/schema/shardingsphere/mode-repository/standalone=org.apache.shardingsphere.spring.namespace.handler.mode.StandaloneModeRepositoryNamespaceHandler
+http\://shardingsphere.apache.org/schema/shardingsphere/mode-repository/standalone=org.apache.shardingsphere.spring.namespace.handler.mode.StandalonePersistRepositoryNamespaceHandler
 http\://shardingsphere.apache.org/schema/shardingsphere/mode-repository/cluster=org.apache.shardingsphere.spring.namespace.handler.mode.ClusterPersistRepositoryNamespaceHandler
